refactor: 참가자와 도착 결과를 사다리 도메인의 하위 도메인으로 판단하고 패키지 변경
