+++ src/main/java/domain/ladder/player/Players.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder.player;
 
 import java.util.Collections;
 import java.util.Iterator;
