+++ src/test/java/domain/GoalTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.ladder.goal.Goal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
