+++ src/test/java/domain/GoalsTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.ladder.goal.Goals;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
