+++ src/test/java/domain/HeightTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.ladder.Height;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
