+++ src/test/java/domain/PlayerTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.ladder.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
